Some more minor naming refactors
